refactor: replace adbpair with automatic pairing, move shared code
