+++ analyst/cmd/adbpair.go
@@ -1,232 +1,220 @@
-// Copyright (c) BARGHEST
-// SPDX-License-Identifier: AGPL-3.0-or-later
-
 package cmd
 
 import (
-	"bufio"
 	"context"
-	"errors"
 	"flag"
 	"fmt"
 	"net"
-	"net/netip"
-	"os"
-	"os/exec"
 	"path/filepath"
-	"regexp"
+	"sort"
 	"strconv"
 	"strings"
+	"sync"
 	"time"
 
 	"github.com/BARGHEST-ngo/androidqf_mesh/acquisition"
 	"github.com/BARGHEST-ngo/androidqf_mesh/adb"
 	"github.com/BARGHEST-ngo/androidqf_mesh/log"
 	"github.com/BARGHEST-ngo/androidqf_mesh/modules"
-	"github.com/BARGHEST-ngo/androidqf_mesh/utils"
 	rt "github.com/botherder/go-savetime/runtime"
 	"github.com/google/uuid"
-	"github.com/mattn/go-isatty"
 	"github.com/peterbourgon/ff/v3/ffcli"
 )
 
-var adbpairArgs struct {
-	host     string
-	hostport string
-	pairport string
-	code     string
-	qf       bool
+type AndroidPeer struct {
+	IP       string
+	HostName string
+	DNSName  string
 }
 
-type Acquisition struct {
-	UUID             string                          `json:"uuid"`
-	AndroidQFVersion string                          `json:"androidqf_version"`
-	StoragePath      string                          `json:"storage_path"`
-	Started          time.Time                       `json:"started"`
-	Completed        time.Time                       `json:"completed"`
-	Collector        *adb.Collector                  `json:"collector"`
-	TmpDir           string                          `json:"tmp_dir"`
-	SdCard           string                          `json:"sdcard"`
-	Cpu              string                          `json:"cpu"`
-	closeLog         func()                          `json:"-"`
-	EncryptedWriter  *acquisition.EncryptedZipWriter `json:"-"`
-	StreamingMode    bool                            `json:"streaming_mode"`
-	StreamingPuller  *acquisition.StreamingPuller    `json:"-"`
+type PairingArgs struct {
+	Host        string
+	PairPort    int
+	DebugPort   int
+	PairingCode string
 }
 
-func AdbpairCmd() *ffcli.Command {
-	fs := flag.NewFlagSet("adbpair", flag.ContinueOnError)
-	fs.StringVar(&adbpairArgs.host, "host", "", "(optional) IP address of the Android device on the MESH network")
-	fs.StringVar(&adbpairArgs.hostport, "hostport", "", "(optional) Wireless Debugging port number")
-	fs.StringVar(&adbpairArgs.pairport, "pairport", "", "(optional) Pairing port from 'Pair device with pairing code' dialog")
-	fs.StringVar(&adbpairArgs.code, "code", "", "(optional) 6-digit pairing code from Android device")
-	fs.BoolVar(&adbpairArgs.qf, "qf", false, "perform adbcollect (AndroidQF/WARD) immediately after connection")
+const maxAutoPairAttempts = 5
+
+var adbpairliteArgs struct {
+	qf bool
+}
+
+func AdbPairCmd() *ffcli.Command {
+	fs := flag.NewFlagSet("adbpairlite", flag.ContinueOnError)
+	fs.BoolVar(&adbpairliteArgs.qf, "qf", false, "perform adbcollect (AndroidQF/WARD) immediately after connection")
 
 	return &ffcli.Command{
-		Name:       "adbpair",
-		ShortUsage: "meshcli adbpair [flags]",
-		ShortHelp:  "Pair & connect to a device on the MESH network via ADB",
-		LongHelp: `The adbpair command initiates an ADB pairing session with a device on the MESH network so you can issue ADB commands. After pairing, it will connect you to the device ready for ADB sessions.
-
-If you run the command without flags, you will be prompted interactively for each required value.
-
-You can use the --qf flag to perform quick forensics immediately on connecting.
-
-Examples:
-  # Interactive mode (recommended for first-time users)
-  meshcli adbpair
-
-  # With all flags specified
-  meshcli adbpair --host 100.64.1.5 --hostport 37891 --pairport 45281 --code 123456
-
-  # Partial flags (will prompt for missing values)
-  meshcli adbpair --host 100.64.1.5
-
-  # Short form with host:port
-  meshcli adbpair 100.64.1.5:37891
-
-  # With quick forensics
-  meshcli adbpair --qf
-
-`,
-		FlagSet: fs,
-		Exec:    runadbpair,
+		Name:       "adbpairlite",
+		ShortUsage: "mesh adbpairlite [flags]",
+		FlagSet:    fs,
+		Exec:       runAdbPair,
 	}
 }
 
-//TODO: adbpairArgs.host should really be defined by the tailscaled network later down the line
-
-func runadbpair(ctx context.Context, args []string) error {
+func runAdbPair(ctx context.Context, args []string) error {
 	ensureAdbConf()
+	pairingArgs := PairingArgs{}
 
-	// For brevity, we also allow `meshcli HOST[:HOSTPORT]`
-	if len(args) == 1 {
-		if strings.Contains(args[0], ":") {
-			var err error
-			adbpairArgs.host, adbpairArgs.hostport, err = net.SplitHostPort(args[0])
-			if err != nil {
-				return fmt.Errorf("failed to parse host:port: %v", err)
+	if len(args) == 0 {
+		fmt.Println("Starting automatic pairing...")
+
+		chosenPeer, err := selectAndroidPeer(ctx)
+		if err != nil {
+			return fmt.Errorf("unable to select an Android client: %w", err)
+		}
+		pairingArgs.Host = chosenPeer.IP
+
+		fmt.Printf("Using Android device: %s (%s)\n\n", chosenPeer.HostName, chosenPeer.IP)
+		fmt.Println("On the Android device:")
+		fmt.Println("1. Enable Wireless Debugging")
+		fmt.Println("2. Tap 'Pair device with pairing code'")
+		ReadString("Press Enter when the pairing dialog is open...")
+
+		fmt.Println("Scanning for open ports...")
+		openPorts, err := scanOpenPorts(chosenPeer.IP)
+		if err != nil {
+			return fmt.Errorf("port scan failed: %w", err)
+		}
+		if len(openPorts) == 0 {
+			return fmt.Errorf("no open ports found on %s - is the pairing dialog open?", chosenPeer.IP)
+		}
+
+		pairingArgs.PairingCode = ReadStringWithValidation("Enter the pairing code shown on the device: ", validatePairingCode)
+
+		adbClient, err := adb.New()
+		if err != nil {
+			return fmt.Errorf("failed to initialize ADB: %w", err)
+		}
+		adb.Client = adbClient
+
+		devices, err := adb.Client.Devices()
+		if err != nil {
+			return fmt.Errorf("failed to get devices: %w", err)
+		}
+		if len(devices) > 0 {
+			fmt.Printf("Found existing ADB devices: %v\n", devices)
+			if err := disconnect(""); err != nil {
+				return err
 			}
-		} else {
-			adbpairArgs.host = args[0]
 		}
-	} else if len(args) > 1 {
-		return fmt.Errorf("unexpected arguments: %v", args)
-	}
 
-	// Check if any arguments are missing and prompt for them interactively
-	if adbpairArgs.host == "" || adbpairArgs.hostport == "" ||
-		adbpairArgs.pairport == "" || adbpairArgs.code == "" {
-		promptForMissingArgs()
-	}
-
-	// Validate all arguments
-	if err := validateAdbArgs(); err != nil {
-		return err
-	}
-
-	adbClient, err := adb.New()
-	if err != nil {
-		return fmt.Errorf("failed to initialize ADB: %v", err)
-	}
-	adb.Client = adbClient
-
-	devices, err := adb.Client.Devices()
-	if err != nil {
-		return fmt.Errorf("failed to get devices: %w", err)
-	}
-	if len(devices) > 0 {
-		fmt.Printf("Found existing ADB devices: %v\n", devices)
-		if err := disconnect(""); err != nil {
+		pairedPort, err := pairWithDiscovery(&pairingArgs, openPorts)
+		if err != nil {
 			return err
 		}
-	}
 
-	if err := adbPair(); err != nil {
-		return err
-	}
-	time.Sleep(2 * time.Second)
-	if err := adbConnect(); err != nil {
-		return err
-	}
+		debugPort, err := resolveDebugPort(chosenPeer.IP, openPorts, pairedPort)
+		if err != nil {
+			return fmt.Errorf("unable to locate debug port: %w", err)
+		}
+		pairingArgs.DebugPort = debugPort
 
-	if adbpairArgs.qf {
-		if err := qf(); err != nil {
+		if err := connect(&pairingArgs); err != nil {
 			return err
 		}
-	}
-	return nil
-}
 
-func validateAdbArgs() error {
-	// Validate host if provided
-	if adbpairArgs.host != "" {
-		if err := validateIPAddress(adbpairArgs.host); err != nil {
-			return fmt.Errorf("invalid --host: %v", err)
+		if err := validateConnect(); err != nil {
+			return err
 		}
-	} else {
-		return errors.New("--host (Host IP) is required")
-	}
 
-	// Validate hostport if provided
-	if adbpairArgs.hostport != "" {
-		if err := validatePort(adbpairArgs.hostport); err != nil {
-			return fmt.Errorf("invalid --hostport: %v", err)
+		if adbpairliteArgs.qf {
+			if err := qf(); err != nil {
+				return err
+			}
 		}
-	} else {
-		return errors.New("--hostport (Wireless Debugging port) is required")
-	}
-
-	// Validate pairport if provided
-	if adbpairArgs.pairport != "" {
-		if err := validatePort(adbpairArgs.pairport); err != nil {
-			return fmt.Errorf("invalid --pairport: %v", err)
-		}
-	} else {
-		return errors.New("--pairport (Pairing port) is required")
-	}
-
-	// Validate code if provided
-	if adbpairArgs.code != "" {
-		if err := validatePairingCode(adbpairArgs.code); err != nil {
-			return fmt.Errorf("invalid --code: %v", err)
-		}
-	} else {
-		return errors.New("--code (Pairing code) is required")
 	}
 
 	return nil
 }
 
-func adbPair() error {
-	checkADBClient()
-	fmt.Printf("Pairing to device...\n")
-	output, err := adb.Client.Exec("pair",
-		fmt.Sprintf("%s:%s", adbpairArgs.host, adbpairArgs.pairport),
-		adbpairArgs.code)
+func pairWithDiscovery(args *PairingArgs, openPorts []int) (int, error) {
+	if len(openPorts) > maxAutoPairAttempts {
+		fmt.Printf("Found %d open ports - too many to try automatically, please identify the pairing port:\n", len(openPorts))
+		choice, ok := promptForPort(openPorts)
+		if !ok {
+			return 0, fmt.Errorf("invalid port selection")
+		}
+		args.PairPort = openPorts[choice]
+		if err := pair(args); err != nil {
+			return 0, err
+		}
+		return args.PairPort, nil
+	}
+
+	for _, port := range openPorts {
+		args.PairPort = port
+		fmt.Printf("Trying port %d...\n", port)
+		if err := pair(args); err != nil {
+			continue
+		}
+		return port, nil
+	}
+	return 0, fmt.Errorf("pairing failed on all discovered ports")
+}
+
+func pair(args *PairingArgs) error {
+	if adb.Client == nil {
+		return fmt.Errorf("adb not initialized")
+	}
+
+	if args == nil {
+		return fmt.Errorf("invalid pairing args")
+	}
+
+	output, err := adb.Client.Exec("pair", net.JoinHostPort(args.Host, strconv.Itoa(args.PairPort)), args.PairingCode)
 	if err != nil {
-		return fmt.Errorf("ADB pair failed: %v\nOutput: %s", err, string(output))
+		return fmt.Errorf("ADB pair failed: %w\nOutput: %s", err, string(output))
 	}
-	fmt.Printf("ADB pair successful:\n%s", string(output))
-	if err := saveHost(adbpairArgs.host); err != nil {
-		return fmt.Errorf("failed to save host config: %v", err)
+	fmt.Println("ADB pair successful")
+	if err := saveHost(args.Host); err != nil {
+		return fmt.Errorf("failed to save host config: %w", err)
 	}
 	return nil
 }
 
-func adbConnect() error {
-	checkADBClient()
-	fmt.Printf("ADB pair successful, now connecting...\n")
-	output, err := adb.Client.Exec("connect",
-		fmt.Sprintf("%s:%s", adbpairArgs.host, adbpairArgs.hostport))
+func resolveDebugPort(ip string, openPorts []int, pairedPort int) (int, error) {
+	var candidates []int
+	for _, p := range openPorts {
+		if p != pairedPort {
+			candidates = append(candidates, p)
+		}
+	}
+
+	switch len(candidates) {
+	case 0:
+		fmt.Println("Scanning for debug port...")
+		return selectPort(ip)
+	case 1:
+		fmt.Printf("Found debug port: %d\n", candidates[0])
+		return candidates[0], nil
+	default:
+		fmt.Println("Multiple debug port candidates, please select:")
+		choice, ok := promptForPort(candidates)
+		if !ok {
+			return 0, fmt.Errorf("invalid port selection")
+		}
+		return candidates[choice], nil
+	}
+}
+
+func connect(args *PairingArgs) error {
+	if adb.Client == nil {
+		return fmt.Errorf("adb not initialized")
+	}
+	if args == nil {
+		return fmt.Errorf("invalid pairing args")
+	}
+	fmt.Printf("Connecting to device...\n")
+	output, err := adb.Client.Exec("connect", net.JoinHostPort(args.Host, strconv.Itoa(args.DebugPort)))
 	if err != nil {
-		return fmt.Errorf("ADB connect failed: %v\nOutput: %s", err, string(output))
+		return fmt.Errorf("ADB connect failed: %w\nOutput: %s", err, string(output))
 	}
-	fmt.Printf("ADB connect successful:\n%s", string(output))
-	if err := saveHostport(adbpairArgs.hostport); err != nil {
-		return fmt.Errorf("failed to save hostport config: %v", err)
+
+	if err := saveHostport(strconv.Itoa(args.DebugPort)); err != nil {
+		return fmt.Errorf("failed to save hostport config: %w", err)
 	}
-	return validateConnect()
+	return nil
 }
 
 func validateConnect() error {
@@ -245,7 +233,7 @@ func validateConnect() error {
 			if strings.HasPrefix(d, "100.") {
 				fmt.Printf("Success! Valid MESH network device\n")
 				fmt.Printf("You may proceed with forensics acquision\n")
-				fmt.Printf("Use ./meshcli adbcollect\n")
+				fmt.Printf("Use mesh adbcollect\n")
 				return nil
 			}
 		}
@@ -255,8 +243,151 @@ func validateConnect() error {
 	}
 }
 
+func selectAndroidPeer(ctx context.Context) (*AndroidPeer, error) {
+	peers, err := getAndroidPeers(ctx)
+	if err != nil || len(peers) == 0 {
+		return nil, fmt.Errorf("unable to find any connected Android clients")
+	}
+
+	var chosenPeer AndroidPeer
+	if len(peers) > 1 {
+		choice, valid := promptForAndroidClient(peers)
+		if !valid {
+			return nil, fmt.Errorf("invalid client selection")
+		}
+		chosenPeer = peers[choice]
+	} else {
+		chosenPeer = peers[0]
+		fmt.Printf("found 1 Android device: %s (%s)\n", chosenPeer.HostName, chosenPeer.IP)
+	}
+
+	return &chosenPeer, nil
+}
+
+func getAndroidPeers(ctx context.Context) ([]AndroidPeer, error) {
+	status, err := localClient.Status(ctx)
+	if err != nil {
+		return nil, fmt.Errorf("failed to get MESH status: %w", err)
+	}
+
+	out := make([]AndroidPeer, 0)
+	for _, p := range status.Peers() {
+		peer := status.Peer[p]
+		if peer.OS != "android" {
+			continue
+		}
+
+		if len(peer.TailscaleIPs) == 0 {
+			continue
+		}
+
+		out = append(out, AndroidPeer{
+			IP:       peer.TailscaleIPs[0].String(),
+			HostName: peer.HostName,
+			DNSName:  peer.DNSName,
+		})
+
+	}
+	return out, nil
+}
+
+func promptForAndroidClient(peers []AndroidPeer) (int, bool) {
+	fmt.Println("Multiple Android clients found, select an Android client:")
+	choices := make([]string, len(peers))
+	for i, p := range peers {
+		choices[i] = fmt.Sprintf("%s (%s)", p.HostName, p.IP)
+	}
+
+	fmt.Print("Select an Android client:\n")
+	return promptForSelection(choices)
+}
+
+func promptForPort(ports []int) (int, bool) {
+	fmt.Println("Multiple open ports found, select a port:")
+	choices := make([]string, len(ports))
+	for i, p := range ports {
+		choices[i] = fmt.Sprintf("%d", p)
+	}
+
+	fmt.Print("Select a port:\n")
+	return promptForSelection(choices)
+}
+
+func promptForSelection(choices []string) (int, bool) {
+	for i, c := range choices {
+		fmt.Printf(" %d) %s \n", i+1, c)
+	}
+
+	var choice int
+	_, err := fmt.Scanln(&choice)
+	if err != nil || choice < 1 || choice > len(choices) {
+		fmt.Println("Invalid selection")
+		return -1, false
+	}
+
+	return choice - 1, true
+}
+
+func selectPort(ip string) (int, error) {
+	fmt.Println("Scanning for port...")
+	openPorts, err := scanOpenPorts(ip)
+	if err != nil {
+		return -1, fmt.Errorf("port scan failed: %w", err)
+	}
+	if len(openPorts) == 0 {
+		return -1, fmt.Errorf("no open ports found on %s", ip)
+	}
+
+	var pairPort int
+	if len(openPorts) == 1 {
+		pairPort = openPorts[0]
+		fmt.Printf("Found port: %d\n", pairPort)
+	} else {
+		choice, ok := promptForPort(openPorts)
+		if !ok {
+			return -1, fmt.Errorf("invalid port selection")
+		}
+
+		pairPort = openPorts[choice]
+	}
+	return pairPort, nil
+}
+
+func scanOpenPorts(ip string) ([]int, error) {
+	const workers = 1000
+	const dialTimeout = 1 * time.Second
+
+	jobs := make(chan int, workers)
+	var mu sync.Mutex
+	var wg sync.WaitGroup
+	var open []int
+
+	for range workers {
+		wg.Go(func() {
+			for p := range jobs {
+				addr := net.JoinHostPort(ip, strconv.Itoa(p))
+				conn, err := net.DialTimeout("tcp", addr, dialTimeout)
+				if err == nil {
+					conn.Close()
+					mu.Lock()
+					open = append(open, p)
+					mu.Unlock()
+				}
+			}
+		})
+	}
+
+	for port := 5000; port <= 65535; port++ {
+		jobs <- port
+	}
+	close(jobs)
+
+	wg.Wait()
+	sort.Ints(open)
+	return open, nil
+}
+
 func qf() error {
-	checkADBClient()
 	devices, err := adb.Client.Devices()
 	if err != nil {
 		return fmt.Errorf("failed to get devices: %w", err)
@@ -264,14 +395,10 @@ func qf() error {
 
 	for _, d := range devices {
 		if strings.HasPrefix(d, "100.") {
-			fmt.Printf("Performing forensics acquisition\n")
-			fmt.Printf("Starting androidqf\n")
+			fmt.Println("Performing forensics acquisition")
+			fmt.Println("Starting androidqf")
 
-			acqt := Acquisition{
-				UUID:             uuid.New().String(),
-				Started:          time.Now().UTC(),
-				AndroidQFVersion: utils.Version,
-			}
+			acqUUID := uuid.New().String()
 
 			for {
 				d, err = adb.Client.SetSerial(d)
@@ -288,39 +415,32 @@ func qf() error {
 				time.Sleep(5 * time.Second)
 			}
 
-			output_folder := filepath.Join(rt.GetExecutableDirectory(), acqt.UUID)
-			acq, err := acquisition.New(output_folder)
+			outputFolder := filepath.Join(rt.GetExecutableDirectory(), acqUUID)
+			acq, err := acquisition.New(outputFolder)
 			if err != nil {
-				log.Debug(err)
 				log.FatalExc("Impossible to initialise the acquisition", err)
 			}
 
 			log.Info(fmt.Sprintf("Started new acquisition in %s", acq.StoragePath))
 
-			mods := modules.List()
-			for _, mod := range mods {
-				err = mod.InitStorage(acq.StoragePath)
-				if err != nil {
+			for _, mod := range modules.List() {
+				if err := mod.InitStorage(acq.StoragePath); err != nil {
 					log.Infof("ERROR: failed to initialize storage for module %s: %v", mod.Name(), err)
 					continue
 				}
-
-				err = mod.Run(acq, false)
-				if err != nil {
+				if err := mod.Run(acq, false); err != nil {
 					log.Infof("ERROR: failed to run module %s: %v", mod.Name(), err)
 				}
 			}
 
-			err = acq.HashFiles()
-			if err != nil {
+			if err := acq.HashFiles(); err != nil {
 				log.ErrorExc("Failed to generate list of file hashes", err)
 				return err
 			}
 
 			acq.StoreInfo()
 
-			err = acq.StoreSecurely()
-			if err != nil {
+			if err := acq.StoreSecurely(); err != nil {
 				log.ErrorExc("Something failed while encrypting the acquisition", err)
 				log.Warning("WARNING: The secure storage of the acquisition folder failed! The data is unencrypted!")
 			}
@@ -331,124 +451,4 @@ func qf() error {
 		}
 	}
 	return nil
-}
-
-func disconnect(serial string) error {
-	checkADBClient()
-	if serial == "" {
-		fmt.Printf("Disconnecting all devices...\n")
-		out, err := exec.Command(adb.Client.ExePath, "disconnect").Output()
-		if err != nil {
-			return fmt.Errorf("failed to disconnect all devices: %v\nOutput: %s", err, string(out))
-		}
-	} else {
-		fmt.Printf("Disconnecting device %s...\n", serial)
-		out, err := exec.Command(adb.Client.ExePath, "disconnect", serial).Output()
-		if err != nil {
-			return fmt.Errorf("failed to disconnect device %s: %v\nOutput: %s", serial, err, string(out))
-		}
-	}
-	return nil
-}
-
-func checkADBClient() {
-	if adb.Client == nil {
-		panic("ADB client not initialized")
-	}
-}
-
-// validateIPAddress validates that the input is a valid IP address (IPv4 or IPv6).
-func validateIPAddress(ip string) error {
-	if ip == "" {
-		return errors.New("IP address cannot be empty")
-	}
-	if _, err := netip.ParseAddr(ip); err != nil {
-		return fmt.Errorf("invalid IP address format")
-	}
-	return nil
-}
-
-// validatePort validates that the input is a valid port number (1-65535).
-func validatePort(port string) error {
-	if port == "" {
-		return errors.New("port cannot be empty")
-	}
-	portNum, err := strconv.ParseUint(port, 10, 16)
-	if err != nil {
-		return errors.New("port must be a number")
-	}
-	if portNum < 1 || portNum > 65535 {
-		return errors.New("port must be between 1 and 65535")
-	}
-	return nil
-}
-
-// validatePairingCode validates that the input is a valid 6-digit pairing code.
-func validatePairingCode(code string) error {
-	if code == "" {
-		return errors.New("pairing code cannot be empty")
-	}
-	// ADB pairing codes are typically 6 digits
-	matched, _ := regexp.MatchString(`^\d{6}$`, code)
-	if !matched {
-		return errors.New("pairing code must be exactly 6 digits")
-	}
-	return nil
-}
-
-// promptForMissingArgs interactively prompts the user for any missing ADB pairing arguments.
-func promptForMissingArgs() {
-	if adbpairArgs.host == "" {
-		adbpairArgs.host = ReadStringWithValidation("Enter target device IP address: ", validateIPAddress)
-	}
-
-	if adbpairArgs.hostport == "" {
-		adbpairArgs.hostport = ReadStringWithValidation("Enter target device Wireless Debugging port: ", validatePort)
-	}
-
-	if adbpairArgs.pairport == "" {
-		adbpairArgs.pairport = ReadStringWithValidation("Enter target device pairing port: ", validatePort)
-	}
-
-	if adbpairArgs.code == "" {
-		adbpairArgs.code = ReadStringWithValidation("Enter pairing code: ", validatePairingCode)
-	}
-}
-
-// ReadString prompts the user for input with the given message and returns the trimmed response.
-// If there is no TTY on both Stdin and Stdout, returns an empty string.
-func ReadString(msg string) string {
-	if !(isatty.IsTerminal(os.Stdin.Fd()) && isatty.IsTerminal(os.Stdout.Fd())) {
-		return ""
-	}
-	fmt.Print(msg)
-	reader := bufio.NewReader(os.Stdin)
-	resp, err := reader.ReadString('\n')
-	if err != nil {
-		return ""
-	}
-	return strings.TrimSpace(resp)
-}
-
-// ValidationFunc is a function that validates input and returns an error if invalid.
-type ValidationFunc func(string) error
-
-// ReadStringWithValidation prompts the user for input with the given message,
-// validates it using the provided validation function, and re-prompts on validation errors.
-// If there is no TTY on both Stdin and Stdout, returns an empty string.
-func ReadStringWithValidation(msg string, validate ValidationFunc) string {
-	if !(isatty.IsTerminal(os.Stdin.Fd()) && isatty.IsTerminal(os.Stdout.Fd())) {
-		return ""
-	}
-	for {
-		input := ReadString(msg)
-		if input == "" {
-			continue
-		}
-		if err := validate(input); err != nil {
-			fmt.Printf("Invalid input: %v\n", err)
-			continue
-		}
-		return input
-	}
 }
